refactor(worker): #113 이미지 처리 패키지 경계 정리
